+++ src/main/java/spot/spot/domain/chat/controller/ChatController.java
@@ -1,0 +1,10 @@
+package spot.spot.domain.chat.controller;
+
+import org.springframework.stereotype.Controller;
+
+import lombok.RequiredArgsConstructor;
+
+@Controller
+@RequiredArgsConstructor
+public class ChatController {
+}
