+++ src/main/java/spot/spot/domain/chat/entity/ChatParticipant.java
@@ -1,4 +1,4 @@
-package spot.spot.domain.chat.messagestatus.entity;
+package spot.spot.domain.chat.entity;
 
 import org.hibernate.annotations.Comment;
 
@@ -13,20 +13,18 @@ import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import spot.spot.domain.chat.chatroom.entity.ChatRoom;
-import spot.spot.domain.chat.message.entity.Message;
 import spot.spot.domain.member.entity.Member;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-@Table(name = "message_status")
-@Comment("메세지 상태")
-public class MessageStatus {
+@Table(name = "chat_participants")
+@Comment("채팅 참여자")
+public class ChatParticipant {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Comment("메세지 상태 아이디")
+	@Comment("채팅 참여자 아이디")
 	private Long id;
 
 	@ManyToOne(fetch = FetchType.LAZY)
@@ -38,12 +36,4 @@ public class MessageStatus {
 	@JoinColumn(name = "member_id", nullable = false)
 	@Comment("회원 아이디")
 	private Member member;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "message_id", nullable = false)
-	@Comment("메세지 아이디")
-	private Message message;
-
-	@Comment("읽음 여부")
-	private boolean isRead;
 }
