+++ src/main/java/spot/spot/domain/chat/entity/ReadStatus.java
@@ -1,4 +1,4 @@
-package spot.spot.domain.chat.chatparticipant.entity;
+package spot.spot.domain.chat.entity;
 
 import org.hibernate.annotations.Comment;
 
@@ -13,19 +13,18 @@ import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import spot.spot.domain.chat.chatroom.entity.ChatRoom;
 import spot.spot.domain.member.entity.Member;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-@Table(name = "chat_participants")
-@Comment("채팅 참여자")
-public class ChatParticipant {
+@Table(name = "message_status")
+@Comment("메세지 상태")
+public class ReadStatus {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Comment("채팅 참여자 아이디")
+	@Comment("메세지 상태 아이디")
 	private Long id;
 
 	@ManyToOne(fetch = FetchType.LAZY)
@@ -37,4 +36,12 @@ public class ChatParticipant {
 	@JoinColumn(name = "member_id", nullable = false)
 	@Comment("회원 아이디")
 	private Member member;
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "message_id", nullable = false)
+	@Comment("메세지 아이디")
+	private ChatMessage message;
+
+	@Comment("읽음 여부")
+	private boolean isRead;
 }
