+++ src/main/java/spot/spot/domain/chat/service/ChatService.java
@@ -1,0 +1,10 @@
+package spot.spot.domain.chat.service;
+
+import org.springframework.stereotype.Service;
+
+import lombok.RequiredArgsConstructor;
+
+@Service
+@RequiredArgsConstructor
+public class ChatService {
+}
